apply comments
