--- index.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const path = require('path');
-const app = express();
-
-app.use(express.static(__dirname+ '/docs'))
-
-// Routes
-app.get('/', function (req, res) {
-    res.sendFile(path.join(__dirname,'docs','index.html'));  
-    
-});
-
-// Port Listen
-app.listen(3000);
-console.log('Running...');
