+++ src/index.html
@@ -1,0 +1,139 @@
+<!DOCTYPE html>
+<html>
+<title>Call Center Team Dashboard</title>
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+  <link rel="stylesheet" href="./styles/style.css" />
+  <script src="https://sdk-cdn.mypurecloud.com/javascript/68.0.0/purecloud-platform-client-v2.min.js"></script>
+  <script src="./scripts/main.js" type="module"></script>
+</head>
+
+
+<body class="w3-light-grey">
+
+  <!-- Top container -->
+  <div class="w3-bar w3-top w3-large">
+    <span class="w3-bar-item w3-right w3-text-white" id="date">Data from  </span>
+  </div>
+
+  <!-- Sidebar/menu -->
+  <nav class="w3-sidebar w3-collapse "><br>
+    <div class="w3-container w3-row">
+      <div class="w3-col s8 w3-bar">
+        <span>Welcome, <strong id="userName"></strong></span><br>
+      </div>
+    </div>
+    <hr>
+    <div class="w3-container">
+      <h5>Dashboard</h5>
+    </div>
+    <div class="w3-bar-block">
+      <a href="#" class="w3-bar-item w3-button w3-padding"><i class="fa fa-users fa-fw"></i>  Overview</a>
+    </div>
+  </nav>
+
+
+  <!-- Overlay effect when opening sidebar on small screens -->
+  <div class="w3-overlay w3-hide-large w3-animate-opacity"></div>
+
+  <!-- !PAGE CONTENT! -->
+  <div class="w3-main">
+
+    <!-- Header -->
+    <header class="w3-container w3-padding-16">
+      <h5><b><i class="fa fa-dashboard"></i> Agents Dashboard</b></h5>
+    </header>
+
+    <div class="w3-row-padding w3-margin-bottom w3-margin-top">
+      <div class="w3-quarter">
+        <div class="w3-container w3-orange w3-padding-16">
+          <div class="w3-left"><i class="fa fa-comment w3-xxxlarge"></i></div>
+          <div class="w3-right">
+            <h3 id="chatInteractionsNumber"></h3>
+          </div>
+          <div class="w3-clear"></div>
+          <h4>Chat Interactions</h4>
+        </div>
+      </div>
+      <div class="w3-quarter">
+        <div class="w3-container w3-navy w3-padding-16">
+          <div class="w3-left"><i class="fa fa-phone-square w3-xxxlarge"></i></div>
+          <div class="w3-right">
+            <h3 id="numberofCalls"></h3>
+          </div>
+          <div class="w3-clear"></div>
+          <h4>Number of Calls</h4>
+        </div>
+      </div>
+      <div class="w3-quarter">
+        <div class="w3-container w3-teal w3-padding-16">
+          <div class="w3-left"><i class="fa fa-users w3-xxxlarge"></i></div>
+          <div class="w3-right">
+            <h3 id="numberofAbandonedCalls"></h3>
+          </div>
+          <div class="w3-clear"></div>
+          <h4>Abandoned Calls</h4>
+        </div>
+      </div>
+      <div class="w3-quarter">
+        <div class="w3-container w3-light-blue w3-text-white w3-padding-16">
+          <div class="w3-left"><i class="fa fa-phone w3-xxxlarge"></i></div>
+          <div class="w3-right">
+            <h3 id="numberofAnsweredCalls"></h3>
+          </div>
+          <div class="w3-clear"></div>
+          <h4>Answered Calls</h4>
+        </div>
+      </div>
+    </div>
+
+    <div class="w3-panel">
+      <div class="w3-row-padding">
+        <div class="w3-twothird">
+          <h5>Activities</h5>
+          <table class="w3-table w3-striped w3-white">
+            <tr>
+              <td><i class="fa fa-user-circle w3-large fa-accent-charcoal"></i></td>
+              <td>Voice Inbound</td>
+              <td><i id="numberofInbound"></i></td>
+            </tr>
+            <tr>
+              <td><i class="fa fa-volume-control-phone w3-large fa-accent-yellow"></i></td>
+              <td>Voice Outbound</td>
+              <td><i id="numberofOutbound"></i></td>
+            </tr>
+            <tr>
+            </tr>
+          </table>
+        </div>
+      </div>
+    </div>
+    <hr>
+    <hr>
+
+    <div class="w3-container">
+      <h6>Select Agent</h6> <select id="agentsList"></select>
+      <h5>Agent Details</h5>
+
+      <table class="w3-table w3-striped w3-bordered w3-border w3-hoverable w3-white" id="userTable">
+        <tr>
+          <td> <b>Start Date and Time</b></td>
+          <td><b>End Date and Time</b></td>
+          <td><b>System Presence</b></td>
+        </tr>
+      </table><br>
+    </div>
+    <hr>
+    <hr>
+
+    <br>
+
+  </div>
+  </div>
+</body>
+
+</html>
