+++ src/package.json
@@ -1,0 +1,7 @@
+{
+  "dependencies": {
+    "bulma": "^0.9.3",
+    "express": "^4.17.1",
+    "purecloud-platform-client-v2": "^124.0.0"
+  }
+}
