+++ src/scripts/config.js
@@ -1,0 +1,16 @@
+
+
+const configuration = {
+    // Token Implicit Grant Client ID
+    clientID: 'your-client-id-here',
+
+    redirectUri: 'http://localhost:3000/',
+
+    genesysCloud: {
+        // Genesys Cloud region
+        // eg. 'mypurecloud.au', 'euw2.pure.cloud', etc...
+        region: 'mypurecloud.com'
+    },
+}
+
+export default configuration
